refactor(opencode): split Agent Swarm session stream
- extract client_config resolution, history transport, and stream event translation from agency-swarm.ts

- update repo maintainability policy to use the 1,200-line cap consistently
